[refactory] route.js 파일에 inputValue 상태 추가
